Added wait times
Control and Trajectory nodes need wait times before starting.
This is because if, simulator is not ready, this nodes will send
messages that will not be caught by the simulator. When the simulator
is finally ready the drone will crash.

The estimator does not need wait time. This is because the only thing
we need is for it to start and stabilize before controller node uses
it.