Angular Velocity Dynamics * Added modelling of angular velocity dynamics as a 2nd order linear system. The identification is done using rise time, settling time, and damping ratio to identify parameters of transfer function in s-domain.
Added some files to control package. This are direct copy of the
files in tests package but with file, topic and node names
according to the convention defined.

Coding Standard

Added a file in which we will keep our coding rules. Please
refer to this file often as it will be probably changing
constantly.